pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amodal-scene-diff"
version = "0.1.0"
description = "Single-view conditioned 3D scene posterior diffusion and posterior selection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "PyYAML>=6.0",
  "numpy>=1.24",
  "Pillow>=10.0",
  "pyarrow>=15.0",
  "trimesh>=4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
